third_party/WebKit/LayoutTests/crypto/aes-ctr-generateKey-parseAlgorithm-failures-expected.txt

CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Missing required property
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
Tests generateKey() with bad AES-CTR parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
